feature(settings): Move settings from package.json to baumeister.json
